TMXMapLoader:
See http://www.badlogicgames.com/forum/viewtopic.php?f=17&t=8613&sid=9a4a4b7368e08f4e285476d48afc694e

This change implements point 1 and other minor changes, and makes point 3 a bit easier to do.
Point 4 was already implemented.

Changes that could affect existing users of the library are marked with "*".

Numbers in square brackets refer to footnotes.

Tile Map Objects:

Loads Tile Map Objects as TextureMapObject instances and populates their texture (previously these were loaded as RectangleMapObject *[1]).

Populates the 'gid' in the properties collection.

Assigns properties of this based on special custom properties in the .TMX file, if present (see the JavaDoc on TmxLoader for the property names).
(e.g. If a property called "rotation" is defined in Tiled, the rotation of this object will be set to it).

I added 'map' as a parameter to the protected method 'loadObject' (I need it and others might too) *[2] .

Added a method than can be overridden to load Image Layers.

Exceptions:

Added the inner exception to some exceptions thrown.
Removed a full stop from the end of some exception messages.

Added some comments (JavaDoc and implementation comments on existing and new code).

Some lines showing as changed are whitespace changes (some the result of running the Eclipse formatting tool on selected lines).

There is a "//TODO" flag in the code, regarding whether the TextureMapObject's opacity should be set there.

-----

[1] This could impact code that uses these already (users of the library), but I think that it is unlikely that they are being used since the current implementation does not provide any way to get the texture nor the 'gid'.
We could avoid this impact by making TextureMapObject extend RectangleMapObject.

[2] This could potentially affect a user who has overridden this method.
An alternative would be to populate the a protected TiledMap instance (maybe the existing 'map' field, if that doesn't break anything), so that it can be used by loading methods.
In fact, I recommend this, and we could do the same with the layer.

- If we don't create protected 'map' and 'layer' references, I think we should add them as parameters to my new methods with names beginning with "loaded".
(And before the next release - before anyone may have written overridding methods in game-specific code).